Used Double where it makes sense.

+++ src/main/kotlin/nl/jamiecraane/raytracing/buildingblocks/Plane.kt
@@ -1,3 +1,3 @@
 package nl.jamiecraane.raytracing.buildingblocks
 
-data class Plane(val normal: Vect3, val D: Float) {}
+data class Plane(val normal: Vect3, val D: Double) {}
